Factorise la modale d'envoi d'email pour une demande
