+++ src/pages/pro/demandes.tsx
@@ -10,6 +10,7 @@ import AdditionalInformation from '@/components/Manager/AdditionalInformation';
 import Comment from '@/components/Manager/Comment';
 import Contact from '@/components/Manager/Contact';
 import Contacted from '@/components/Manager/Contacted';
+import DemandEmailForm from '@/components/Manager/DemandEmailForm';
 import DemandStatusBadge from '@/components/Manager/DemandStatusBadge';
 import Status from '@/components/Manager/Status';
 import Tag from '@/components/Manager/Tag';
@@ -23,6 +24,7 @@ import Icon from '@/components/ui/Icon';
 import Indicator from '@/components/ui/Indicator';
 import Link from '@/components/ui/Link';
 import Loader from '@/components/ui/Loader';
+import ModalSimple from '@/components/ui/ModalSimple';
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from '@/components/ui/Resizable';
 import TableSimple, { type ColumnDef, type QuickFilterPreset } from '@/components/ui/TableSimple';
 import Tooltip from '@/components/ui/Tooltip';
@@ -50,212 +52,6 @@ const displayModeDeChauffage = (demand: Demand) => {
   }
   return demand['Type de chauffage'];
 };
-
-function getDemandsTableColumns(updateDemand: (demandId: string, demandUpdate: Partial<Demand>) => Promise<void>): ColumnDef<Demand>[] {
-  return [
-    {
-      id: 'indicators',
-      header: '',
-      align: 'center',
-      cell: ({ row }) => (
-        <div className="flex flex-col gap-2">
-          {row.original.Status === DEMANDE_STATUS.EMPTY && !row.original['Prise de contact'] && (
-            <Tooltip title="Prospect non recontacté et statut en attente de prise en charge">
-              <Icon name="fr-icon-flag-fill" size="sm" color="red" />
-            </Tooltip>
-          )}
-          {row.original.haut_potentiel && <Badge type="haut_potentiel" />}
-        </div>
-      ),
-      width: '46px',
-    },
-    {
-      accessorKey: 'Status',
-      header: 'Statut',
-      cell: ({ row }) => <Status demand={row.original} updateDemand={updateDemand} />,
-      width: '290px',
-      filterType: 'Facets',
-      filterProps: {
-        Component: ({ value }) => <DemandStatusBadge status={value as DemandStatus} />,
-      },
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Prise de contact',
-      header: 'Prospect recontacté',
-      cell: ({ row }) => <Contacted demand={row.original} updateDemand={updateDemand} />,
-      align: 'center',
-      filterType: 'Facets',
-      width: '85px',
-      enableGlobalFilter: false,
-    },
-    {
-      accessorFn: (row) => `${row.Nom} ${row.Prénom} ${row.Mail}`,
-      header: 'Contact',
-      cell: ({ row }) => <Contact demand={row.original} updateDemand={updateDemand} />,
-      width: '280px',
-      enableSorting: false,
-    },
-    {
-      accessorKey: 'Adresse',
-      header: () => (
-        <div className="flex items-center">
-          Adresse
-          <Tooltip
-            iconProps={{
-              className: 'ml-1',
-            }}
-            title="La mention 'PDP' est indiquée pour les adresses situées dans le périmètre de développement prioritaire d'un réseau classé (connu par France Chaleur Urbaine)."
-          />
-        </div>
-      ),
-      cell: ({ row }) => (
-        <div className="whitespace-normal">
-          {row.original.Adresse}
-          {row.original['en PDP'] === 'Oui' && <Badge type="pdp" />}
-        </div>
-      ),
-      width: '220px',
-      enableSorting: false,
-    },
-    {
-      accessorKey: 'Date de la demande',
-      header: 'Date de la demande',
-      cellType: 'Date',
-      width: '94px',
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Structure',
-      header: 'Type',
-      cell: ({ row }) => <Tag text={row.original.Structure} />,
-      width: '130px',
-      filterType: 'Facets',
-      enableGlobalFilter: false,
-    },
-    {
-      accessorFn: (row) => displayModeDeChauffage(row),
-      header: 'Mode de chauffage',
-      cell: ({ row }) => <Tag text={displayModeDeChauffage(row.original)} />,
-      width: '126px',
-      filterType: 'Facets',
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Distance au réseau',
-      header: () => (
-        <div className="flex items-center">
-          Distance au réseau (m)
-          <Tooltip
-            iconProps={{
-              className: 'ml-1',
-            }}
-            title="Distance à vol d'oiseau"
-          />
-        </div>
-      ),
-      cell: ({ row }) => (
-        <AdditionalInformation demand={row.original} field="Distance au réseau" updateDemand={updateDemand} type="number" />
-      ),
-      width: '120px',
-      filterType: 'Range',
-      filterProps: {
-        domain: [0, 1000],
-        unit: 'm',
-      },
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Identifiant réseau',
-      header: 'ID réseau le plus proche',
-      width: '85px',
-      filterType: 'Facets',
-    },
-    {
-      accessorKey: 'Nom réseau',
-      header: 'Nom du réseau le plus proche',
-      cell: ({ row }) => <div className="whitespace-normal">{row.original['Nom réseau']}</div>,
-      width: '200px',
-    },
-    {
-      accessorKey: 'Logement',
-      header: 'Nb logements (lots)',
-      cell: ({ row }) => <AdditionalInformation demand={row.original} field="Logement" updateDemand={updateDemand} type="number" />,
-      width: '120px',
-      filterType: 'Range',
-      sorting: 'nullsLast',
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Surface en m2',
-      header: 'Surface en m2',
-      cell: ({ row }) => <AdditionalInformation demand={row.original} field="Surface en m2" updateDemand={updateDemand} type="number" />,
-      width: '120px',
-      filterType: 'Range',
-      filterProps: {
-        unit: 'm2',
-      },
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Conso',
-      header: 'Conso gaz (MWh)',
-      cell: ({ row }) => <AdditionalInformation demand={row.original} field="Conso" updateDemand={updateDemand} type="number" />,
-      width: '120px',
-      filterType: 'Range',
-      filterProps: {
-        unit: 'MWh',
-      },
-      enableGlobalFilter: false,
-    },
-    {
-      accessorKey: 'Commentaires',
-      header: 'Commentaires',
-      cell: ({ row }) => <Comment demand={row.original} updateDemand={updateDemand} />,
-      width: '280px',
-      enableSorting: false,
-    },
-    {
-      accessorKey: 'Affecté à',
-      header: () => (
-        <div className="flex items-center">
-          Affecté à
-          <Tooltip
-            iconProps={{
-              className: 'ml-1',
-            }}
-            title={
-              <>
-                "Non affecté" : demande éloignée du réseau non transmise aux opérateurs
-                <br />
-                <br />
-                Vous pouvez ajouter ou modifier une affectation : le changement sera effectif après validation manuelle par l'équipe FCU.
-              </>
-            }
-          />
-        </div>
-      ),
-      cell: ({ row }) => (
-        <AdditionalInformation demand={row.original} field="Affecté à" updateDemand={updateDemand} type="text" width={125} />
-      ),
-      width: '150px',
-      enableSorting: false,
-      filterType: 'Facets',
-    },
-
-    // obligatoire afin d'être utilisables dans les presets
-    {
-      accessorKey: 'haut_potentiel',
-      filterType: 'Facets', // obligatoire pour faire fonctionner le filtre
-      visible: false,
-    },
-    {
-      accessorKey: 'en PDP',
-      filterType: 'Facets', // obligatoire pour faire fonctionner le filtre
-      visible: false,
-    },
-  ];
-}
 
 const quickFilterPresets = {
   all: {
@@ -332,7 +128,7 @@ function DemandesNew(): React.ReactElement {
   const virtualizerRef = useRef<Virtualizer<HTMLDivElement, Element>>(null) as RefObject<Virtualizer<HTMLDivElement, Element>>;
 
   const [selectedDemandId, setSelectedDemandId] = useState<string | null>(null);
-
+  const [modalDemandId, setModalDemandId] = useState<string | null>(null);
   const tableRowSelection = useMemo(() => {
     return selectedDemandId ? { [selectedDemandId]: true } : {};
   }, [selectedDemandId]);
@@ -385,7 +181,212 @@ function DemandesNew(): React.ReactElement {
     );
   }, []);
 
-  const tableColumns = useMemo(() => getDemandsTableColumns(updateDemand), [updateDemand]);
+  const tableColumns: ColumnDef<Demand>[] = useMemo(
+    () => [
+      {
+        id: 'indicators',
+        header: '',
+        align: 'center',
+        cell: ({ row }) => (
+          <div className="flex flex-col gap-2">
+            {row.original.Status === DEMANDE_STATUS.EMPTY && !row.original['Prise de contact'] && (
+              <Tooltip title="Prospect non recontacté et statut en attente de prise en charge">
+                <Icon name="fr-icon-flag-fill" size="sm" color="red" />
+              </Tooltip>
+            )}
+            {row.original.haut_potentiel && <Badge type="haut_potentiel" />}
+          </div>
+        ),
+        width: '46px',
+      },
+      {
+        accessorKey: 'Status',
+        header: 'Statut',
+        cell: ({ row }) => <Status demand={row.original} updateDemand={updateDemand} />,
+        width: '290px',
+        filterType: 'Facets',
+        filterProps: {
+          Component: ({ value }) => <DemandStatusBadge status={value as DemandStatus} />,
+        },
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Prise de contact',
+        header: 'Prospect recontacté',
+        cell: ({ row }) => <Contacted demand={row.original} updateDemand={updateDemand} />,
+        align: 'center',
+        filterType: 'Facets',
+        width: '85px',
+        enableGlobalFilter: false,
+      },
+      {
+        accessorFn: (row) => `${row.Nom} ${row.Prénom} ${row.Mail}`,
+        header: 'Contact',
+        cell: ({ row }) => <Contact demand={row.original} onEmailClick={setModalDemandId} />,
+        width: '280px',
+        enableSorting: false,
+      },
+      {
+        accessorKey: 'Adresse',
+        header: () => (
+          <div className="flex items-center">
+            Adresse
+            <Tooltip
+              iconProps={{
+                className: 'ml-1',
+              }}
+              title="La mention 'PDP' est indiquée pour les adresses situées dans le périmètre de développement prioritaire d'un réseau classé (connu par France Chaleur Urbaine)."
+            />
+          </div>
+        ),
+        cell: ({ row }) => (
+          <div className="whitespace-normal">
+            {row.original.Adresse}
+            {row.original['en PDP'] === 'Oui' && <Badge type="pdp" />}
+          </div>
+        ),
+        width: '220px',
+        enableSorting: false,
+      },
+      {
+        accessorKey: 'Date de la demande',
+        header: 'Date de la demande',
+        cellType: 'Date',
+        width: '94px',
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Structure',
+        header: 'Type',
+        cell: ({ row }) => <Tag text={row.original.Structure} />,
+        width: '130px',
+        filterType: 'Facets',
+        enableGlobalFilter: false,
+      },
+      {
+        accessorFn: (row) => displayModeDeChauffage(row),
+        header: 'Mode de chauffage',
+        cell: ({ row }) => <Tag text={displayModeDeChauffage(row.original)} />,
+        width: '126px',
+        filterType: 'Facets',
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Distance au réseau',
+        header: () => (
+          <div className="flex items-center">
+            Distance au réseau (m)
+            <Tooltip
+              iconProps={{
+                className: 'ml-1',
+              }}
+              title="Distance à vol d'oiseau"
+            />
+          </div>
+        ),
+        cell: ({ row }) => (
+          <AdditionalInformation demand={row.original} field="Distance au réseau" updateDemand={updateDemand} type="number" />
+        ),
+        width: '120px',
+        filterType: 'Range',
+        filterProps: {
+          domain: [0, 1000],
+          unit: 'm',
+        },
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Identifiant réseau',
+        header: 'ID réseau le plus proche',
+        width: '85px',
+        filterType: 'Facets',
+      },
+      {
+        accessorKey: 'Nom réseau',
+        header: 'Nom du réseau le plus proche',
+        cell: ({ row }) => <div className="whitespace-normal">{row.original['Nom réseau']}</div>,
+        width: '200px',
+      },
+      {
+        accessorKey: 'Logement',
+        header: 'Nb logements (lots)',
+        cell: ({ row }) => <AdditionalInformation demand={row.original} field="Logement" updateDemand={updateDemand} type="number" />,
+        width: '120px',
+        filterType: 'Range',
+        sorting: 'nullsLast',
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Surface en m2',
+        header: 'Surface en m2',
+        cell: ({ row }) => <AdditionalInformation demand={row.original} field="Surface en m2" updateDemand={updateDemand} type="number" />,
+        width: '120px',
+        filterType: 'Range',
+        filterProps: {
+          unit: 'm2',
+        },
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Conso',
+        header: 'Conso gaz (MWh)',
+        cell: ({ row }) => <AdditionalInformation demand={row.original} field="Conso" updateDemand={updateDemand} type="number" />,
+        width: '120px',
+        filterType: 'Range',
+        filterProps: {
+          unit: 'MWh',
+        },
+        enableGlobalFilter: false,
+      },
+      {
+        accessorKey: 'Commentaires',
+        header: 'Commentaires',
+        cell: ({ row }) => <Comment demand={row.original} updateDemand={updateDemand} />,
+        width: '280px',
+        enableSorting: false,
+      },
+      {
+        accessorKey: 'Affecté à',
+        header: () => (
+          <div className="flex items-center">
+            Affecté à
+            <Tooltip
+              iconProps={{
+                className: 'ml-1',
+              }}
+              title={
+                <>
+                  "Non affecté" : demande éloignée du réseau non transmise aux opérateurs
+                  <br />
+                  <br />
+                  Vous pouvez ajouter ou modifier une affectation : le changement sera effectif après validation manuelle par l'équipe FCU.
+                </>
+              }
+            />
+          </div>
+        ),
+        cell: ({ row }) => (
+          <AdditionalInformation demand={row.original} field="Affecté à" updateDemand={updateDemand} type="text" width={125} />
+        ),
+        width: '150px',
+        enableSorting: false,
+        filterType: 'Facets',
+      },
+
+      // obligatoire afin d'être utilisables dans les presets
+      {
+        accessorKey: 'haut_potentiel',
+        filterType: 'Facets', // obligatoire pour faire fonctionner le filtre
+        visible: false,
+      },
+      {
+        accessorKey: 'en PDP',
+        filterType: 'Facets', // obligatoire pour faire fonctionner le filtre
+        visible: false,
+      },
+    ],
+    [updateDemand]
+  );
 
   const onFeatureClick = useCallback(
     (feature: MapGeoJSONFeature) => {
@@ -446,6 +447,7 @@ function DemandesNew(): React.ReactElement {
       ) && columnFilters.length === preset.filters.length
     );
   };
+  const modalDemand = demands.find((demand) => demand.id === modalDemandId);
 
   return (
     <SimplePage
@@ -453,6 +455,14 @@ function DemandesNew(): React.ReactElement {
       description="Votre tableau de bord pour la gestion des demandes des réseaux de chaleur"
       mode="authenticated"
     >
+      <ModalSimple
+        title={`Envoi d'un courriel à ${modalDemand?.Mail}`}
+        open={!!modalDemandId}
+        size="large"
+        onOpenChange={(open) => !open && setModalDemandId(null)}
+      >
+        {modalDemand && <DemandEmailForm currentDemand={modalDemand} updateDemand={updateDemand} />}
+      </ModalSimple>
       <div className="mb-8">
         <div className="flex items-center flex-wrap">
           <Input
@@ -539,4 +549,4 @@ function DemandesNew(): React.ReactElement {
 
 export default DemandesNew;
 
-export const getServerSideProps = withAuthentication(['gestionnaire', 'demo']);
+export const getServerSideProps = withAuthentication(['gestionnaire', 'demo', 'admin']);
